requirements.txt
Flask
Flask-SQLAlchemy
pdfplumber
python-dotenv
Flask-Cors
PyJWT
Flask-Bcrypt
PyMuPDF
openai
